appi/root  api setup moved to main
