+++ src/main.rs
@@ -7,9 +7,14 @@ pub mod query_engine;
 pub mod service;
 pub mod shared;
 
+use axum::{routing::post, Extension, Router};
 use error::AppError;
-use api::start_server;
+use tokio::net::TcpListener;
+use tower::ServiceBuilder;
+use tower_http::trace::TraceLayer;
 use tracing_subscriber::EnvFilter;
+
+use crate::service::QueryService;
 
 #[tokio::main]
 async fn main() -> Result<(), AppError> {
@@ -17,7 +22,20 @@ async fn main() -> Result<(), AppError> {
         .with_env_filter(EnvFilter::from_default_env())
         .init();
 
-    start_server(config::QUERY_SERVER).await;
+    let query_srv = QueryService::new();
+
+    let app = Router::new()
+        .route("/", axum::routing::get(api::root_handler))
+        .route("/query", post(api::query_handler))
+        .layer(Extension(query_srv));
+
+    let listener = TcpListener::bind(config::QUERY_SERVER).await.unwrap();
+
+    let middleware_stack = ServiceBuilder::new()
+        .layer(TraceLayer::new_for_http())
+        .into_inner();
+
+    axum::serve(listener, app.layer(middleware_stack)).await.unwrap();
 
     Ok(())
 }
